+++ shaders/lib/sunrays_vol.glsl
@@ -52,7 +52,7 @@ float getVolSunraysAmount(float depth, inout uint rng  ARGS_OUT) {
 	#include "/import/eyeBrightnessSmooth.glsl"
 	float skyBrightness = eyeBrightnessSmooth.y / 240.0;
 	sunraysAmount += (sunraysAmount > 0.0 ? 1.0 : 0.0) * mix(SUNRAYS_MIN_UNDERGROUND, SUNRAYS_MIN_SURFACE, skyBrightness);
-	sunraysAmount *= 0.005;
+	sunraysAmount *= 0.01;
 	
 	#include "/import/gbufferModelViewInverse.glsl"
 	vec3 playerPos = (gbufferModelViewInverse * startMat(viewPos)).xyz;